Rename ebmdatalab module to gcutils
