+++ openprescribing/frontend/management/commands/convert_hscic_prescribing.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 from django.core.management.base import CommandError
 
-from ebmdatalab.bigquery import Client, TableExporter
+from gcutils.bigquery import Client, TableExporter
 
 logger = logging.getLogger(__name__)
