+++ openprescribing/gcutils/table_dumper.py
